+++ src/ClockifyCli/Models/NewProject.cs
@@ -1,0 +1,5 @@
+using Newtonsoft.Json;
+
+namespace ClockifyCli.Models;
+
+public record NewProject([property: JsonProperty("name")] string Name);